Complete endpoints and npm scripts
