+++ package.json
@@ -7,7 +7,8 @@
     "test": "npm test",
     "start": "./node_modules/.bin/pm2 start ./build/index.js --name nmp-project",
     "stop": "./node_modules/.bin/pm2 stop nmp-project",
-    "start-dev": "nodemon index.js"
+    "start-dev": "nodemon index.js",
+    "build": "tsc"
   },
   "author": "J Alejandro Duarte S",
   "license": "MIT",
