+++ src/cart/cart.controller.ts
@@ -1,10 +1,33 @@
-import express, { Express, Handler, Router } from 'express';
+import express, { Express, Handler, Router, NextFunction, Request, Response } from 'express';
+import Joi from 'joi';
 import { CartBase, ProductBase, UpdateCartRequestBody } from '../types';
 import { CartService } from './cart.service';
+import { calculateTotal } from '../utils/utils';
 
-const CartController = (cartRepository: CartBase, productRepository: ProductBase, middleware: Handler) : Router => {
+const CartController = (cartRepository: CartBase, productRepository: ProductBase) : Router => {
   const cartRouter: Router = express.Router();
   const cartService: CartService = new CartService(cartRepository);
+
+  const putCartSchema = Joi.object({
+    productId: Joi.string().uuid().required(),
+    count: Joi.number().integer().min(0).required()
+  });
+
+  const validatePutCartBody = async (req: Request, res: Response, next: NextFunction) => {
+    try {
+      console.log(req.body);
+      await putCartSchema.validateAsync(req.body);
+      next();
+    } catch (err) {
+      console.log("logging there", err)
+      res.status(400).send({
+        "data": null,
+        "error": {
+          "message": "Products are not valid"
+        }
+      });
+    }
+  }
 
   cartRouter.get('/cart', (req, res, next) => {
     const userId = req.get('x-user-id');
@@ -12,12 +35,18 @@ const CartController = (cartRepository: CartBase, productRepository: ProductBase
     res
       .status(200)
       .send({
-        "data": cart,
+        "data": {
+          "cart": {
+            "cart": cart.id,
+            "items": cart.items
+          },
+          "total": calculateTotal(cart.items)
+        },
         "error": null
       });
   });
 
-  cartRouter.put('/cart', middleware ,(req, res, next) => {
+  cartRouter.put('/cart', validatePutCartBody ,(req, res, next) => {
     const { productId, count } = req.body as UpdateCartRequestBody;
     const userId = req.get('x-user-id') as string;
     try {
@@ -26,7 +55,13 @@ const CartController = (cartRepository: CartBase, productRepository: ProductBase
       res
         .status(200)
         .send({
-          "data": cart,
+          "data": {
+            "cart": {
+              "cart": cart.id,
+              "items": cart.items
+            },
+            "total": calculateTotal(cart.items)
+          },
           "error": null
         });
     } catch (error) {
